aider: The `impute_column` function in `regmean_imputer\impute_column.py` has been modified to ensure that the number of splits for KFold cross-validation is not greater than the number of samples in the `non_missing_data` DataFrame.
# Aider chat conversation:

USER: `test_impute_column` fails with ValueError: Cannot have number of splits n_splits=5 greater than the number of samples n_samples=4

ASSISTANT: The error is due to the fact that the number of splits for KFold cross-validation in the `impute_column` function is greater than the number of samples in the `non_missing_data` DataFrame.

To fix this, we can add a check in the `impute_column` function to ensure that the number of splits is not greater than the number of samples. If it is, we can set the number of splits to be equal to the number of samples.

Here are the changes needed:

1. Modify the `impute_column` function in the `regmean_imputer\impute_column.py` file to add a check for the number of splits.

Here is the *edit block* for this change:

```python
regmean_imputer\impute_column.py
<<<<<<< HEAD
    kf = KFold(n_splits=n_splits, shuffle=True)
=======
    # Ensure that the number of splits is not greater than the number of samples
    n_splits = min(n_splits, len(non_missing_data))
    kf = KFold(n_splits=n_splits, shuffle=True)
>>>>>>> updated
```

Please confirm this change.